refactor:
- use class serializer interceptor globally as it will be used in every module
