+++ src/modules/tags/tags.controller.ts
@@ -1,10 +1,11 @@
 import { Controller, Get, Query } from '@nestjs/common';
 import { TagsService } from './tags.service';
 import { PaginationDto } from 'src/common/dto';
-import { sendSuccessResponse } from 'src/utils/response.handler';
-import { Public } from '../auth/ParamDecorator';
+import { sendSuccessResponse } from 'src/utils/response-handler/success.response-handler';
+import { ApiTags } from '@nestjs/swagger';
 
 @Controller('tags')
+@ApiTags('Tags')
 export class TagsController {
   constructor(private readonly tagsService: TagsService) {}
